CSQC is standard now, and secure via the same md4 as a map currently has. Particle system functions renamed a bit, a few other cleanups in that area. Console handling tweeked. Better rules for subconsoles and plugins. Commands are coloured if it'll be execed, which should help reduce occurences of chat being commands. tab compleation tweeked, partial compleation no longer changes the suggestion.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@895 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/qclib/qcc_pr_comp.c
@@ -7254,7 +7254,7 @@ void QCC_PR_ParseDefs (char *classname)
 			extern int ForcedCRC;
 			if (ForcedCRC)
 				QCC_PR_ParseError(ERR_BADEXTENSION, "progs crc was already specified - only one is allowed");
-			ForcedCRC = pr_immediate._float;
+			ForcedCRC = (int)pr_immediate._float;
 			QCC_PR_Lex();
 			QCC_PR_Expect(";");
 			return;
